(refactor/feat) Break query helpers into separate file; add addClick func
